TrueFoundry/examples/requirements.txt
fastapi
uvicorn
zscaler-sdk-python
requests
python-dotenv
